📦 chore(docker-compose.dev.yml): add docker-compose file for development environment
This commit adds a new file `docker-compose.dev.yml` which contains the configuration for running the application in a development environment using Docker Compose. The file defines the following services:

- `backend`: The backend service for the application. It builds the backend code and runs it on port 3001. It also sets environment variables for the application, such as the database URL and JWT secret.
- `db-studio`: A service for running the database studio. It builds the backend code and runs the database studio on port 5555. It depends on the `backend` service.
- `db`: The PostgreSQL database service. It uses the latest PostgreSQL image and runs on port 5432. It sets environment variables for the database username and password.

The services are connected to an `app-network` bridge network.

Note: The commented out `frontend` service is not included in this commit.